Decoupled Architecure
